services/service1/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn==0.24.0

# Base de datos
pymongo==4.6.0
redis==5.0.1

# Variables de entorno
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2

# Validación de datos
pydantic==2.5.0

# Certificados SSL
certifi==2023.11.17

# CORS (ya incluido en FastAPI pero explícito)
# fastapi ya incluye starlette que maneja CORS